+++ package.json
@@ -16,6 +16,7 @@
     ]
   },
   "dependencies": {
+    "@react-native-async-storage/async-storage": "^1.19.5",
     "expo": "~49.0.15",
     "expo-status-bar": "~1.6.0",
     "jest-expo": "^49.0.0",
